+++ src/roles/grades/b4.ts
@@ -1,0 +1,5 @@
+import CustomRole from "../../types/customRole";
+import makeGradeRole from "../../utils/makeGradeRole";
+
+const b4Role: CustomRole = makeGradeRole({ grade: "B4", color: [0, 112, 255], reason: "B4 Grade Role" });
+export default b4Role;
