+++ src/roles/grades/d2.ts
@@ -1,0 +1,5 @@
+import CustomRole from "../../types/customRole";
+import makeGradeRole from "../../utils/makeGradeRole";
+
+const d2Role: CustomRole = makeGradeRole({ grade: "D2", color: [112, 48, 160], reason: "D2 Grade Role" });
+export default d2Role;
