+++ src/roles/grades/m1.ts
@@ -1,0 +1,5 @@
+import CustomRole from "../../types/customRole";
+import makeGradeRole from "../../utils/makeGradeRole";
+
+const m1Role: CustomRole = makeGradeRole({ grade: "M1", color: [0, 176, 80], reason: "M1 Grade Role" });
+export default m1Role;
