+++ src/roles/grades/obog.ts
@@ -1,0 +1,6 @@
+import CustomRole from "../../types/customRole";
+import makeGradeRole from "../../utils/makeGradeRole";
+
+const obogRole: CustomRole = makeGradeRole({ grade: "OB/OG", color: [128, 128, 128], reason: "OB/OG Grade Role" });
+
+export default obogRole;
